+++ site/src/content/docs/user-guide/concepts/tools/vended-tools-imports.ts
@@ -50,3 +50,13 @@ import { makeSleep } from '@strands-agents/sdk/vended-tools/sleep'
 import { Agent } from '@strands-agents/sdk'
 import { stop } from '@strands-agents/sdk/experimental/vended-tools/stop'
 // --8<-- [end:stop_import]
+
+// --8<-- [start:web_fetch_import]
+import { Agent } from '@strands-agents/sdk'
+import { webFetch } from '@strands-agents/sdk/vended-tools/web-fetch'
+// --8<-- [end:web_fetch_import]
+
+// --8<-- [start:web_fetch_markdown_import]
+import { Agent } from '@strands-agents/sdk'
+import { makeWebFetch } from '@strands-agents/sdk/vended-tools/web-fetch'
+// --8<-- [end:web_fetch_markdown_import]
