+++ site/src/content/docs/user-guide/concepts/tools/vended-tools.ts
@@ -8,6 +8,7 @@ import { notebook } from '@strands-agents/sdk/vended-tools/notebook'
 import { SessionManager, FileStorage } from '@strands-agents/sdk'
 import { sleep, makeSleep } from '@strands-agents/sdk/vended-tools/sleep'
 import { stop } from '@strands-agents/sdk/experimental/vended-tools/stop'
+import { webFetch, makeWebFetch } from '@strands-agents/sdk/vended-tools/web-fetch'
 
 // Agent with vended tools example
 async function agentWithVendedToolsExample() {
@@ -161,4 +162,21 @@ async function stopExample() {
   })
   await agent.invoke('Summarize the changes in ./CHANGELOG.md')
   // --8<-- [end:stop_example]
+}
+
+// Web fetch example
+async function webFetchExample() {
+  // --8<-- [start:web_fetch_example]
+  const agent = new Agent({ tools: [webFetch] })
+  await agent.invoke('Summarize https://example.com/blog/post')
+  // --8<-- [end:web_fetch_example]
+}
+
+// Web fetch markdown mode example
+async function webFetchMarkdownExample() {
+  // --8<-- [start:web_fetch_markdown_example]
+  const fetch = makeWebFetch({ mode: 'markdown' })
+  const agent = new Agent({ tools: [fetch] })
+  await agent.invoke('Read https://example.com/docs and explain the architecture')
+  // --8<-- [end:web_fetch_markdown_example]
 }
